+++ src/main/java/com/autonomouslogic/everef/cli/publishrefdata/PublishRefData.java
@@ -2,6 +2,8 @@ package com.autonomouslogic.everef.cli.publishrefdata;
 
 import com.autonomouslogic.commons.rxjava3.Rx3Util;
 import com.autonomouslogic.everef.cli.Command;
+import com.autonomouslogic.everef.cli.publishrefdata.bundle.GroupBundleRenderer;
+import com.autonomouslogic.everef.cli.publishrefdata.bundle.TypeBundleRenderer;
 import com.autonomouslogic.everef.config.Configs;
 import com.autonomouslogic.everef.http.OkHttpHelper;
 import com.autonomouslogic.everef.model.ReferenceEntry;
