+++ src/main/java/com/autonomouslogic/everef/cli/publishrefdata/bundle/TypeBundleRenderer.java
@@ -1,12 +1,9 @@
-package com.autonomouslogic.everef.cli.publishrefdata;
+package com.autonomouslogic.everef.cli.publishrefdata.bundle;
 
 import com.autonomouslogic.everef.refdata.InventoryType;
 import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.node.ObjectNode;
 import io.reactivex.rxjava3.core.Flowable;
 import io.reactivex.rxjava3.core.Maybe;
-import java.util.Optional;
-import java.util.function.Function;
 import javax.inject.Inject;
 import lombok.extern.log4j.Log4j2;
 import org.apache.commons.lang3.tuple.Pair;
@@ -41,8 +38,8 @@ public class TypeBundleRenderer extends BundleRenderer {
 		bundleTraits(type, typesJson);
 		bundleVariations(type, typesJson);
 		bundleRequiredSkills(type, skillsJson, typesJson);
-		bundleUnits(unitsJson, attributesJson);
-		bundleIcons(iconsJson, attributesJson);
+		bundleDogmaAttributesUnits(attributesJson, unitsJson);
+		bundleDogmaAttributesIcons(attributesJson, iconsJson);
 
 		if (!attributesJson.isEmpty()) {
 			bundleJson.set("dogma_attributes", attributesJson);
@@ -59,20 +56,5 @@ public class TypeBundleRenderer extends BundleRenderer {
 
 		var path = refDataUtil.subPath("types", type.getTypeId()) + "/bundle";
 		return Maybe.just(Pair.of(path, bundleJson));
-	}
-
-	private void bundleTraits(InventoryType type, ObjectNode typesJson) {
-		Optional.ofNullable(type.getTraits())
-				.flatMap(traits -> Optional.ofNullable(traits.getTypes()))
-				.map(types -> types.keySet().stream())
-				.stream()
-				.flatMap(Function.identity())
-				.map(Long::parseLong)
-				.forEach(typeId -> {
-					var json = getTypesMap().get(typeId);
-					if (json != null) {
-						typesJson.set(Long.toString(typeId), json);
-					}
-				});
 	}
 }
